Создал отдельный docker-compose.local.yaml (для dev среды) * добавил example_env.env * поправил Dockerfile * поправил poetry.lock * вынес schemas.py * вынес constants.py * добавил TODO и комментарии
