+++ .dockerignore
@@ -1,0 +1,2 @@
+poetry.lock
+venv
