+++ Dockerfile
@@ -1,8 +1,9 @@
 FROM python:3.11.5
 
 RUN pip install poetry==1.5.1
-COPY pyproject.toml poetry.lock ./
+COPY pyproject.toml ./
 RUN poetry install
 
-COPY community_management community_management
+RUN pwd
+COPY . ./
 ENTRYPOINT ["uvicorn", "--host", "0.0.0.0", "--port", "8000", "community_management.app:app"]
