+++ community_management/community/constants.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class TopicKind(str, Enum):
+    FUN = "fun"
+    MOVIES = "movies"
+    SCIENCE = "science"
+    SPORTS = "sports"
