+++ docker-compose.local.yaml
@@ -15,12 +15,11 @@ services:
       timeout: 10s
       retries: 3
   db:
-    # Указат явно тег для image (postgres)
-    image: postgres
+    image: postgres:15
     environment:
       POSTGRES_USER: admin
       POSTGRES_PASSWORD: admin
-      POSTGRES_DB: admin
+      POSTGRES_DB: db
     ports:
       - "5432:5432"
     volumes:
