+++ example_env.env
@@ -1,0 +1,9 @@
+# Здесь надо поместить переменные для DB, Redis и прочих
+POSTGRES_USER=admin
+POSTGRES_PASSWORD=admin
+POSTGRES_DB=db
+
+REDIS_HOST=
+REDIS_PORT=
+# обязательно ставить в production
+REDIS_PASSWORD=
